backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pandas
neo4j
torch
torch-geometric
shap
pydantic
python-dotenv